GPencil: Integrate perimter functions
This is a modified version of the functions included in D6637

https://developer.blender.org/D6637

+++ source/blender/io/gpencil/intern/gpencil_io_base.h
@@ -45,7 +45,7 @@ class GpencilExporter {
   bool gpencil_3d_point_to_screen_space(struct ARegion *region,
                                         const float diff_mat[4][4],
                                         const float co[3],
-                                        int r_co[2]);
+                                        float r_co[2]);
 
   std::string rgb_to_hex(float color[3]);
   std::string to_lower_string(char *input_text);
